+++ source/renderer/app/stores/StakingStore.js
@@ -5,7 +5,11 @@ import { orderBy, take } from 'lodash';
 import Store from './lib/Store';
 import Request from './lib/LocalizedRequest';
 import { ROUTES } from '../routes-config';
-import { RECENT_STAKE_POOLS_COUNT } from '../config/stakingConfig';
+import {
+  RECENT_STAKE_POOLS_COUNT,
+  STAKE_POOLS_INTERVAL,
+  STAKE_POOLS_FAST_INTERVAL,
+} from '../config/stakingConfig';
 import type {
   Reward,
   RewardForIncentivizedTestnet,
@@ -17,10 +21,9 @@ import StakePool from '../domains/StakePool';
 import REWARDS from '../config/stakingRewards.dummy.json';
 
 export default class StakingStore extends Store {
-  STAKE_POOLS_INITIAL_INTERVAL = 1000; // 1000 milliseconds
-  STAKE_POOLS_REFRESH_INTERVAL = 30 * 60 * 1000; // 30 minutes | unit: milliseconds;
+  @observable fetchingStakePoolsFailed = false;
 
-  initialPooling: ?IntervalID = null;
+  pollingStakePoolsInterval: ?IntervalID = null;
   refreshPooling: ?IntervalID = null;
 
   startDateTime: string = '2019-12-09T00:00:00.161Z';
@@ -29,9 +32,12 @@ export default class StakingStore extends Store {
   percentage: number = 14;
 
   setup() {
-    this.initialPooling = setInterval(
-      this.refreshStakePoolsData,
-      this.STAKE_POOLS_INITIAL_INTERVAL
+    // Initial fetch
+    this.getStakePoolsData();
+    // Set fetch interval to 30 minutes
+    this.pollingStakePoolsInterval = setInterval(
+      this.getStakePoolsData,
+      STAKE_POOLS_INTERVAL
     );
     const { staking } = this.actions;
     staking.goToStakingInfoPage.listen(this._goToStakingInfoPage);
@@ -39,7 +45,6 @@ export default class StakingStore extends Store {
       this._goToStakingDelegationCenterPage
     );
     staking.joinStakePool.listen(this._joinStakePool);
-    this.refreshStakePoolsData();
   }
 
   // REQUESTS
@@ -118,19 +123,112 @@ export default class StakingStore extends Store {
     return new Date(this.startDateTime).getTime() - new Date().getTime() > 0;
   }
 
-  @action refreshStakePoolsData = async () => {
-    const { isSynced, isConnected } = this.stores.networkStatus;
-    if (this.stores.wallets._pollingBlocked || !isSynced || !isConnected)
+  @action getStakePoolsData = async () => {
+    const { stores } = this;
+    const { networkStatus, wallets } = stores;
+    const { isSynced, isConnected } = networkStatus;
+    const { _pollingBlocked } = wallets;
+
+    if (
+      (_pollingBlocked || !isSynced || !isConnected) &&
+      !this.refreshPooling
+    ) {
+      this._resetPolling(true);
       return;
-    if (this.initialPooling && !this.refreshPooling) {
-      clearInterval(this.initialPooling);
-      this.initialPooling = null;
-      this.refreshPooling = setInterval(
-        this.refreshStakePoolsData,
-        this.STAKE_POOLS_REFRESH_INTERVAL
-      );
     }
-    await this.stakePoolsRequest.execute().promise;
+
+    try {
+      await this.stakePoolsRequest.execute().promise;
+      if (this.refreshPooling) this._resetPolling(false);
+    } catch (error) {
+      if (!this.refreshPooling) {
+        this._resetPolling(true);
+      }
+    }
+  };
+
+  @action _resetPolling = fetchFailed => {
+    if (fetchFailed) {
+      this.fetchingStakePoolsFailed = true;
+      clearInterval(this.pollingStakePoolsInterval);
+      this.pollingStakePoolsInterval = null;
+      this.refreshPooling = setInterval(
+        this.getStakePoolsData,
+        STAKE_POOLS_FAST_INTERVAL
+      );
+    } else {
+      this.fetchingStakePoolsFailed = false;
+      clearInterval(this.refreshPooling);
+      this.refreshPooling = null;
+      if (!this.pollingStakePoolsInterval) {
+        this.pollingStakePoolsInterval = setInterval(
+          this.getStakePoolsData,
+          STAKE_POOLS_INTERVAL
+        );
+      }
+    }
+  };
+
+  // For testing only
+  @action _setFakePoller = forceLoading => {
+    const { stores, environment } = this;
+    const { networkStatus, wallets } = stores;
+    const { isSynced, isConnected } = networkStatus;
+    const { _pollingBlocked } = wallets;
+
+    // Enable faker only for development node (NODE_ENV = 'development')
+    if (environment.isDev) {
+      if (forceLoading) {
+        // Reset all staking pollers
+        if (this.refreshPooling) {
+          clearInterval(this.refreshPooling);
+          this.refreshPooling = null;
+        }
+        if (this.pollingStakePoolsInterval) {
+          clearInterval(this.pollingStakePoolsInterval);
+          this.pollingStakePoolsInterval = null;
+        }
+        this.fetchingStakePoolsFailed = true;
+        return;
+      }
+
+      // Regular fetching way with faked response that throws error.
+      if (
+        (_pollingBlocked || !isSynced || !isConnected) &&
+        !this.refreshPooling
+      ) {
+        this._resetPolling(true);
+        return;
+      }
+
+      try {
+        throw new Error('Faked "Stake pools" fetch error');
+      } catch (error) {
+        if (!this.refreshPooling) {
+          this._resetPolling(true);
+        }
+      }
+    }
+  };
+
+  // For testing only
+  @action _setFakedStakePools = () => {
+    if (this.environment.isDev) {
+      if (this.refreshPooling) {
+        clearInterval(this.refreshPooling);
+        this.refreshPooling = null;
+      }
+      if (this.pollingStakePoolsInterval) {
+        clearInterval(this.pollingStakePoolsInterval);
+        this.pollingStakePoolsInterval = null;
+      }
+      const newStakePools = [
+        this.stakePoolsRequest.result[1],
+        this.stakePoolsRequest.result[2],
+      ];
+      this.stakePoolsRequest.reset();
+      this.stakePoolsRequest.result = newStakePools;
+    }
   };
 
   _goToStakingInfoPage = () => {